feat: Added support for stack traces by default, updated README, updated CI/Dependabot config
