+++ .github/dependabot.yml
@@ -18,3 +18,5 @@ updates:
       - "major"
   schedule:
     interval: "weekly"
+  cooldown:
+      default-days: 5
